% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa_utils.R
\name{csa_utils}
\alias{csa_utils}
\title{Cross-Sectional Average (CSA) Utilities}
\description{
Internal functions for computing cross-sectional averages and their lags,
used to approximate unobserved common factors in CCE-type estimators.
}
\keyword{internal}
