% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorb.R
\name{.absorb_demean_one}
\alias{.absorb_demean_one}
\title{Within transform: demean numeric columns by a single factor}
\usage{
.absorb_demean_one(X, g)
}
\arguments{
\item{X}{Numeric vector or matrix.}

\item{g}{A factor of length matching the rows of X.}
}
\value{
Demeaned X.
}
\description{
Within transform: demean numeric columns by a single factor
}
\keyword{internal}
