% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_rcce.R
\name{.ahn_horenstein_er}
\alias{.ahn_horenstein_er}
\title{Ahn-Horenstein ER criterion}
\usage{
.ahn_horenstein_er(eigenvalues, K_max)
}
\arguments{
\item{eigenvalues}{Numeric vector of eigenvalues (decreasing order).}

\item{K_max}{Integer: maximum number of factors to consider.}
}
\value{
Integer: estimated number of factors.
}
\description{
Eigenvalue Ratio criterion: \verb{k* = argmax_\{k=1,...,K_max\} lambda[k] / lambda[k+1]}
}
\keyword{internal}
