% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_rcce.R
\name{.ahn_horenstein_gr}
\alias{.ahn_horenstein_gr}
\title{Ahn-Horenstein GR criterion}
\usage{
.ahn_horenstein_gr(eigenvalues, K_max)
}
\arguments{
\item{eigenvalues}{Numeric vector of eigenvalues.}

\item{K_max}{Integer: maximum number of factors.}
}
\value{
Integer: estimated number of factors.
}
\description{
Growth Ratio criterion:
\code{GR(k) = log(1 + lambda[k]/V_k) / log(1 + lambda[k+1]/V_k)}
where \code{V_k = sum(lambda[k+1:end])}.
}
\keyword{internal}
