% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_exp.R
\name{.bkp_residual}
\alias{.bkp_residual}
\title{BKP (2019) residual method for alpha}
\usage{
.bkp_residual(em, N, T_val, n_pca, tuning, n_bootstrap)
}
\description{
BKP (2019) residual method for alpha
}
\keyword{internal}
