% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_exp.R
\name{.bkp_variable}
\alias{.bkp_variable}
\title{BKP (2016) variable method for alpha}
\usage{
.bkp_variable(em, N, T_val, n_pca, test_size)
}
\description{
BKP (2016) variable method for alpha
}
\keyword{internal}
