% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd_test.R
\name{.cd_star}
\alias{.cd_star}
\title{CD* (Pesaran & Xie 2021) bias-corrected statistic}
\usage{
.cd_star(rho_ij, N, T_val, em, n_pca)
}
\description{
CD* (Pesaran & Xie 2021) bias-corrected statistic
}
\keyword{internal}
