% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{.compute_irf}
\alias{.compute_irf}
\title{Compute IRF from AR coefficients and impulse coefficient}
\usage{
.compute_irf(ar_coefs, ar_orders, max_p, beta_impulse, H)
}
\description{
Compute IRF from AR coefficients and impulse coefficient
}
\keyword{internal}
