% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dcce.R
\name{.estimate_half}
\alias{.estimate_half}
\title{Estimate MG on a subset of time periods}
\usage{
.estimate_half(
  panel,
  y_name,
  x_names,
  csa_colnames,
  unit_var,
  time_var,
  constant,
  trend,
  time_subset
)
}
\description{
Estimate MG on a subset of time periods
}
\keyword{internal}
