% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ife.R
\name{.ife_select_factors}
\alias{.ife_select_factors}
\title{Select number of factors via BIC3 (Bai & Ng 2002)}
\usage{
.ife_select_factors(Y_dm, max_r = 10L)
}
\description{
Select number of factors via BIC3 (Bai & Ng 2002)
}
\keyword{internal}
