% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_break.R
\name{.make_segments}
\alias{.make_segments}
\title{Split a time vector into segments based on break dates}
\usage{
.make_segments(all_times, break_dates)
}
\description{
Split a time vector into segments based on break dates
}
\keyword{internal}
