% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd_test.R
\name{.pairwise_correlations}
\alias{.pairwise_correlations}
\title{Compute pairwise correlations of residuals}
\usage{
.pairwise_correlations(em)
}
\description{
Compute pairwise correlations of residuals
}
\keyword{internal}
