% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cce.R
\name{.partial_out}
\alias{.partial_out}
\title{Partial out CSAs from regressors}
\usage{
.partial_out(X, Z)
}
\arguments{
\item{X}{Numeric matrix of variables to be partialled.}

\item{Z}{Numeric matrix of CSA regressors (including intercept if present).}
}
\value{
Numeric matrix of residuals after partialling out Z.
}
\description{
Projects out the CSA (cross-sectional average) component from a matrix
of variables using the Frisch-Waugh-Lovell projection:
\verb{M_Z X = X - Z (Z'Z)^\{-1\} Z' X}
}
\keyword{internal}
