% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_utils.R
\name{.pinv}
\alias{.pinv}
\title{Moore-Penrose pseudoinverse}
\usage{
.pinv(M, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{M}{A numeric matrix.}

\item{tol}{Tolerance for zero singular values.}
}
\value{
The pseudoinverse of \code{M}.
}
\description{
Moore-Penrose pseudoinverse
}
\keyword{internal}
