% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{.print_coef_table}
\alias{.print_coef_table}
\title{Internal: print a single coefficient table}
\usage{
.print_coef_table(b, se)
}
\description{
Internal: print a single coefficient table
}
\keyword{internal}
