% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects.R
\name{.register_marginaleffects_methods}
\alias{.register_marginaleffects_methods}
\title{Register marginaleffects S3 methods at package load}
\usage{
.register_marginaleffects_methods()
}
\description{
Called from \code{.onLoad()} in \code{R/zzz.R}. Dynamically registers
methods on the \pkg{marginaleffects} generics if that package is
installed. Safe no-op otherwise.
}
\keyword{internal}
