% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd_test.R
\name{.resid_vector_to_matrix}
\alias{.resid_vector_to_matrix}
\title{Convert stacked residual vector to matrix}
\usage{
.resid_vector_to_matrix(resids, panel)
}
\description{
Convert stacked residual vector to matrix
}
\keyword{internal}
