% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_break.R
\name{.seq_bai_perron}
\alias{.seq_bai_perron}
\title{Sequential Bai-Perron: recursively search pre/post segments}
\usage{
.seq_bai_perron(
  panel,
  formula,
  model,
  unit_index,
  time_index,
  first_break,
  trim,
  n_breaks,
  test_terms,
  ...
)
}
\description{
Sequential Bai-Perron: recursively search pre/post segments
}
\keyword{internal}
