% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegration.R
\name{.validate_westerlund_tests}
\alias{.validate_westerlund_tests}
\title{Validate Westerlund test names}
\usage{
.validate_westerlund_tests(test)
}
\description{
Validate Westerlund test names
}
\keyword{internal}
