% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{.var_to_matrix}
\alias{.var_to_matrix}
\title{Internal: convert a single variable column of a panel to a (N x T) matrix}
\usage{
.var_to_matrix(panel, var)
}
\description{
Internal: convert a single variable column of a panel to a (N x T) matrix
}
\keyword{internal}
