% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegration.R
\name{.westerlund_aggregate}
\alias{.westerlund_aggregate}
\title{Aggregate unit-level ECM statistics into Westerlund Ga/Gt/Pa/Pt}
\usage{
.westerlund_aggregate(unit_res, tests = c("ga", "gt", "pa", "pt"))
}
\description{
Aggregate unit-level ECM statistics into Westerlund Ga/Gt/Pa/Pt
}
\keyword{internal}
