% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{fitted.dcce_fit}
\alias{fitted.dcce_fit}
\title{Extract fitted values from a dcce_fit object}
\usage{
\method{fitted}{dcce_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
A numeric vector of fitted values (y-hat), or \code{NULL} if not
available.
}
\description{
Extract fitted values from a dcce_fit object
}
