% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{glance.dcce_fit}
\alias{glance.dcce_fit}
\title{Glance at a dcce_fit object}
\usage{
\method{glance}{dcce_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
A single-row tibble.
}
\description{
Returns a single-row tibble of model summary statistics, compatible with
the broom package.
}
