% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_tests.R
\name{hausman_test}
\alias{hausman_test}
\title{Hausman-type Test: MG vs Pooled}
\usage{
hausman_test(object)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}
}
\value{
An object of class \code{dcce_hausman} with the test statistic,
degrees of freedom, and p-value.
}
\description{
Tests the null that the pooled and MG estimators are both consistent
(i.e. slopes are homogeneous and the pooled estimator is efficient)
against the alternative that slopes are heterogeneous and only MG is
consistent. The test statistic is
\deqn{H = (\hat\beta_{MG} - \hat\beta_{pool})'
          [V_{MG} - V_{pool}]^{-1}
          (\hat\beta_{MG} - \hat\beta_{pool}) \sim \chi^2_k.}
}
