% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Information Criteria for CSA Selection}
\usage{
ic(object, models = NULL)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{models}{Optional list of \code{dcce_fit} objects for PC criteria comparison.}
}
\value{
An object of class \code{dcce_ic} with \code{IC1}, \code{IC2}, \code{PC1}, \code{PC2}.
}
\description{
Implements the Margaritella & Westerlund (2023) information and panel
criteria for selecting the number of cross-sectional averages.
}
\keyword{internal}
