% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{plot.dcce_irf}
\alias{plot.dcce_irf}
\title{Plot a dcce_irf object}
\usage{
\method{plot}{dcce_irf}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_irf} object.}

\item{...}{Passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot a dcce_irf object
}
