% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{predict.dcce_fit}
\alias{predict.dcce_fit}
\title{Predict from a dcce_fit object}
\usage{
\method{predict}{dcce_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
A numeric vector of in-sample predictions.
}
\description{
Predict from a dcce_fit object
}
