% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.dcce_boot}
\alias{print.dcce_boot}
\title{Print method for dcce_boot objects}
\usage{
\method{print}{dcce_boot}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_boot} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for dcce_boot objects
}
