% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_break.R
\name{print.dcce_break}
\alias{print.dcce_break}
\title{Print a dcce_break object}
\usage{
\method{print}{dcce_break}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_break} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_break object
}
