% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd_test.R
\name{print.dcce_cd}
\alias{print.dcce_cd}
\title{Print method for dcce_cd objects}
\usage{
\method{print}{dcce_cd}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_cd} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for dcce_cd objects
}
