% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cips.R
\name{print.dcce_cips}
\alias{print.dcce_cips}
\title{Print a dcce_cips object}
\usage{
\method{print}{dcce_cips}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_cips} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_cips object
}
