% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegration.R
\name{print.dcce_cointegration}
\alias{print.dcce_cointegration}
\title{Print method for dcce_cointegration}
\usage{
\method{print}{dcce_cointegration}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_cointegration} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for dcce_cointegration
}
