% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegration_extra.R
\name{print.dcce_cointegration_extra}
\alias{print.dcce_cointegration_extra}
\title{Print a dcce_cointegration_extra object}
\usage{
\method{print}{dcce_cointegration_extra}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_cointegration_extra} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_cointegration_extra object
}
