% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_exp.R
\name{print.dcce_csd}
\alias{print.dcce_csd}
\title{Print method for dcce_csd objects}
\usage{
\method{print}{dcce_csd}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_csd} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for dcce_csd objects
}
