% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{print.dcce_fit}
\alias{print.dcce_fit}
\title{Print a dcce_fit object}
\usage{
\method{print}{dcce_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_fit object
}
