% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger.R
\name{print.dcce_granger}
\alias{print.dcce_granger}
\title{Print a dcce_granger object}
\usage{
\method{print}{dcce_granger}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_granger} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_granger object
}
