% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_tests.R
\name{print.dcce_hausman}
\alias{print.dcce_hausman}
\title{Print a dcce_hausman object}
\usage{
\method{print}{dcce_hausman}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_hausman} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_hausman object
}
