% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{print.dcce_irf}
\alias{print.dcce_irf}
\title{Print a dcce_irf object}
\usage{
\method{print}{dcce_irf}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_irf} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_irf object
}
