% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{print.dcce_rolling}
\alias{print.dcce_rolling}
\title{Print a dcce_rolling object}
\usage{
\method{print}{dcce_rolling}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_rolling} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_rolling object
}
