% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_tests.R
\name{print.dcce_swamy}
\alias{print.dcce_swamy}
\title{Print a dcce_swamy object}
\usage{
\method{print}{dcce_swamy}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_swamy} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_swamy object
}
