% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_root_extra.R
\name{print.dcce_unit_root}
\alias{print.dcce_unit_root}
\title{Print a dcce_unit_root object}
\usage{
\method{print}{dcce_unit_root}(x, ...)
}
\arguments{
\item{x}{A \code{dcce_unit_root} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcce_unit_root object
}
