% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwt8}
\alias{pwt8}
\title{Penn World Tables Growth Panel Dataset}
\format{
A data frame with 4464 rows and 8 columns:
\describe{
\item{id}{Country numeric identifier}
\item{year}{Year (1960--2007)}
\item{log_rgdpo}{Log real GDP (output-side)}
\item{log_hc}{Log human capital index}
\item{log_ck}{Log physical capital stock}
\item{log_ngd}{Log of population growth plus depreciation}
\item{country}{Country identifier (character)}
\item{d_log_rgdpo}{First difference of log_rgdpo (GDP growth)}
}
}
\source{
Downloaded from
\url{https://github.com/JanDitzen/xtdcce2}.
}
\usage{
pwt8
}
\description{
Panel dataset from Jan Ditzen's xtdcce2 Stata package, originally
derived from the Penn World Tables 8. Used in Ditzen (2018, Stata
Journal 18:3, 585--617) to illustrate MG, CCE, and DCCE estimation.
Contains 93 countries observed from 1960 to 2007 (balanced).
}
\references{
Ditzen, J. (2018). Estimating dynamic common-correlated effects in
Stata. \emph{The Stata Journal}, 18(3), 585--617.
}
\keyword{datasets}
