% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_rcce.R
\name{rcce_estimator}
\alias{rcce_estimator}
\title{Regularized CCE Estimator Internals}
\description{
Internal functions for the regularized CCE (rCCE) estimator of Juodis (2022).
Uses SVD of cross-sectional averages to extract principal components,
with the number of components selected by Ahn-Horenstein ER/GR criteria.
}
\keyword{internal}
