% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{residuals.dcce_fit}
\alias{residuals.dcce_fit}
\title{Extract residuals from a dcce_fit object}
\usage{
\method{residuals}{dcce_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
A numeric vector.
}
\description{
Extract residuals from a dcce_fit object
}
