% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{summary.dcce_fit}
\alias{summary.dcce_fit}
\title{Summary for a dcce_fit object}
\usage{
\method{summary}{dcce_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary for a dcce_fit object
}
