% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcce_fit.R
\name{vcov.dcce_fit}
\alias{vcov.dcce_fit}
\title{Extract variance-covariance matrix from a dcce_fit object}
\usage{
\method{vcov}{dcce_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dcce_fit} object.}

\item{...}{Ignored.}
}
\value{
A variance-covariance matrix.
}
\description{
Extract variance-covariance matrix from a dcce_fit object
}
