% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_csdl.R
\name{csdl_estimator}
\alias{csdl_estimator}
\title{CS-DL Estimator Internals}
\description{
Internal functions for the Cross-Sectionally augmented Distributed Lag
(CS-DL) estimator of Chudik et al. (2016). Long-run coefficient \code{w2}
is estimated directly as the coefficient on the level of x.
}
\keyword{internal}
